requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.2.0
torch==2.2.0
ultralytics==8.3.118
supervision==0.18.0 